requirements.txt
PyTelegramBotAPI==4.21.0
openai==1.35.15
requests
python-dotenv==0.19.1
pytest